refactor: move controller models to new package
