+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/AchTransfersController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.ach_transfer.AchTransfer;
 import com.mx.models.ach_transfer.options.AchTransferListOptions;
-import com.mx.web.mdx.models.AchTransfers.AchTransferListQueryParameters;
+import com.mx.path.model.mdx.web.model.ach_transfer.AchTransferListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
