+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/CrossAccountTransferFeesController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.cross_account_transfer.options.FeeListOptions;
 import com.mx.models.transfer.Fee;
-import com.mx.web.mdx.models.CrossAccountTransfers.CrossAccountTransferFeeListQueryParameters;
+import com.mx.path.model.mdx.web.model.cross_account_transfer.CrossAccountTransferFeeListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
