+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/RecurringTransfersController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.transfer.RecurringTransfer;
 import com.mx.models.transfer.options.RecurringTransferListOptions;
-import com.mx.web.mdx.models.Transfers.RecurringTransferListQueryParameters;
+import com.mx.path.model.mdx.web.model.transfer.RecurringTransferListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
