+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/TransferAccountsController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.account.Account;
 import com.mx.models.transfer.options.AccountListOptions;
-import com.mx.web.mdx.models.Transfers.TransferAccountListQueryParameters;
+import com.mx.path.model.mdx.web.model.transfer.TransferAccountListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
