+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/TransferAmountOptionsController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.transfer.TransferAmountOption;
 import com.mx.models.transfer.options.TransferAmountOptionListOptions;
-import com.mx.web.mdx.models.Transfers.TransferAmountOptionsListQueryParameters;
+import com.mx.path.model.mdx.web.model.transfer.TransferAmountOptionsListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
