+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/controller/TransfersController.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.AccessorResponse;
 import com.mx.common.models.MdxList;
 import com.mx.models.transfer.Transfer;
 import com.mx.models.transfer.options.TransferListOptions;
-import com.mx.web.mdx.models.Transfers.TransferListQueryParameters;
+import com.mx.path.model.mdx.web.model.transfer.TransferListQueryParameters;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
