+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/model/ach_transfer/AchScheduledTransferFrequencyListQueryParameters.java
@@ -1,9 +1,9 @@
-package com.mx.web.mdx.models.Transfers;
+package com.mx.path.model.mdx.web.model.ach_transfer;
 
 import lombok.Data;
 
 @SuppressWarnings({ "checkstyle:MemberName", "checkstyle:ParameterName", "checkstyle:MethodName" })
 @Data
-public class RecurringTransferListQueryParameters {
+public class AchScheduledTransferFrequencyListQueryParameters {
   private String transfer_type;
 }
