+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/model/ach_transfer/AchTransferAchAccountListQueryParameters.java
@@ -1,13 +1,14 @@
-package com.mx.web.mdx.models.Transfers;
+package com.mx.path.model.mdx.web.model.ach_transfer;
 
 import lombok.Data;
 
 @SuppressWarnings({ "checkstyle:MemberName", "checkstyle:ParameterName", "checkstyle:MethodName" })
 @Data
-public class TransferAmountOptionsListQueryParameters {
+public class AchTransferAchAccountListQueryParameters {
 
   // NOTE: Non-standard naming to assist with parameter binding
   private String transfer_type;
-  private String source_account_id;
-  private String destination_account_id;
+  private String list_type;
+  private String account_id;
+  private String achAccount_id;
 }
