+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/model/check_image/CheckImagesGetQueryParameters.java
@@ -1,4 +1,4 @@
-package com.mx.web.mdx.models.CheckImages;
+package com.mx.path.model.mdx.web.model.check_image;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
