+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/model/transfer/TransferAmountOptionsListQueryParameters.java
@@ -1,14 +1,13 @@
-package com.mx.web.mdx.models.AchTransfers;
+package com.mx.path.model.mdx.web.model.transfer;
 
 import lombok.Data;
 
 @SuppressWarnings({ "checkstyle:MemberName", "checkstyle:ParameterName", "checkstyle:MethodName" })
 @Data
-public class AchTransferAchAccountListQueryParameters {
+public class TransferAmountOptionsListQueryParameters {
 
   // NOTE: Non-standard naming to assist with parameter binding
   private String transfer_type;
-  private String list_type;
-  private String account_id;
-  private String achAccount_id;
+  private String source_account_id;
+  private String destination_account_id;
 }
