+++ mdx-web/src/main/java/com/mx/path/model/mdx/web/model/transfer/TransferListQueryParameters.java
@@ -1,9 +1,9 @@
-package com.mx.web.mdx.models.AchTransfers;
+package com.mx.path.model.mdx.web.model.transfer;
 
 import lombok.Data;
 
 @SuppressWarnings({ "checkstyle:MemberName", "checkstyle:ParameterName", "checkstyle:MethodName" })
 @Data
-public class AchTransferListQueryParameters {
+public class TransferListQueryParameters {
   private String transfer_type;
 }
