+++ mdx-web/src/test/groovy/com/mx/path/model/mdx/web/controller/TransferRepaymentsControllerTest.groovy
@@ -12,10 +12,8 @@ import com.mx.models.transfer.options.RepaymentListOptions
 import com.mx.path.gateway.api.Gateway
 import com.mx.path.gateway.api.transfer.RepaymentGateway
 import com.mx.path.gateway.api.transfer.TransferGateway
-import com.mx.path.model.mdx.web.controller.BaseController
-import com.mx.path.model.mdx.web.controller.TransferRepaymentsController
+import com.mx.path.model.mdx.web.model.transfer.TransferRepaymentsQueryParameters
 import com.mx.path.testing.WithMockery
-import com.mx.web.mdx.models.Transfers.TransferRepaymentsQueryParameters
 
 import org.mockito.ArgumentCaptor
 import org.mockito.Mockito
