+++ mdx-web/src/test/groovy/com/mx/path/model/mdx/web/controller/TransfersControllerTest.groovy
@@ -10,9 +10,7 @@ import com.mx.models.transfer.Transfer
 import com.mx.models.transfer.options.TransferListOptions
 import com.mx.path.gateway.api.Gateway
 import com.mx.path.gateway.api.transfer.TransferGateway
-import com.mx.path.model.mdx.web.controller.BaseController
-import com.mx.path.model.mdx.web.controller.TransfersController
-import com.mx.web.mdx.models.Transfers.TransferListQueryParameters
+import com.mx.path.model.mdx.web.model.transfer.TransferListQueryParameters
 
 import org.mockito.ArgumentCaptor
 import org.mockito.Mockito
